// src/models/customers_info.model.ts
export interface CUSTOMERS_INFO  {
    customers_info_id?:  number;
    customers_info_date_of_last_logon:  string;
    customers_info_number_of_logons:  number;
    customers_info_date_account_created:  string;
    customers_info_date_account_last_modified:  string;
    global_product_notifications:  number
}